refactoring of Sketck class and creation of BaseSketck class allowing access to Sketck information in the package 'core'
